Rename module
Fix PR issues
