+++ hr_payroll_commission/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+#
+#  File: __init__.py
+#  Module: hr_payroll_commission
+#
+#  Created by sge@open-net.ch
+#  See __openerp__.py for full copyright and licensing details
+##############################################################################
+
+from . import models
