+++ hr_payroll_commission/models/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+#
+#  File: models/__init__.py
+#  Module: hr_payroll_commission
+#
+#  Created by sge@open-net.ch
+#  See __openerp__.py for full copyright and licensing details
+##############################################################################
+
+from . import hr_contract
+from . import hr_expenses
+from . import hr_payslip
+from . import invoices
+from . import account_move_line
