+++ hr_payroll_commission/models/hr_expenses.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 #
-#  File: __openerp__.py
-#  Module: hr_payroll_extended
+#  File: models/hr_expenses.py
+#  Module: hr_payroll_commission
 #
-#  Created by sge@open-net.ch
+#  Created by cyp@open-net.ch
 #
-#  Copyright (c) 2014-TODAY Open-Net Ltd. <http://www.open-net.ch>
+#  Copyright (c) 2015-TODAY Open-Net Ltd.
 ##############################################################################
 #
 #    OpenERP, Open Source Management Solution
-#    Copyright (C) 2004-TODAY OpenERP S.A. <http://www.openerp.com>
+#    Copyright (C) 2004-2010 Tiny SPRL (<http://tiny.be>).
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -27,29 +27,12 @@
 ##############################################################################
 
 
-{
-    'name': 'Extended - Payroll',
-    'summary': 'Payroll Extended',
-    'category': 'Human Resources',
-    'author': "Open-Net Sàrl,Odoo Community Association (OCA)",
-    'depends': [
-        'decimal_precision',
-        'hr_payroll',
-        'hr_payroll_account',
-        'hr_contract',
-        'hr_attendance',
-        'account'
-    ],
-    'version': '9.0.1.0',
-    'auto_install': False,
-    'demo': [],
-    'website': 'http://open-net.ch',
-    'license': 'AGPL-3',
-    'data': [
-        'views/hr_contract_view.xml',
-        'views/hr_payroll_view.xml',
-        'data/hr.salary.rule.category.xml',
-        'data/hr.salary.rule.xml',
-    ],
-    'installable': True
-}
+from openerp import models, fields
+
+
+class HrExpenseExpense(models.Model):
+    _inherit = 'hr.expense'
+
+    # ---------- Fields management
+
+    slip_id = fields.Many2one('hr.payslip', string='Pay slip')
