+++ hr_payroll_commission/views/hr_payroll_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
     File: views/hr_payroll_view.xml
-    Module: hr_payroll_extended
+    Module: hr_payroll_commission
 
     Created by cyp@open-net.ch
